Revert "SILGen: avoid reusing the same opened archetype in keypath setter and getter functions."
There are some crashes (for which we don't have a reproducer, unfortunately), which are caused by this change.

The good thing is that this change is not needed anymore, because it's handled by the more general eecb9fa975c0a83d9cadf0429bcb15a80a6f06b6 "SILModule: track opened archetypes per function.".

The test case for this was added in the original commit (test/stdlib/KeyPath.swift), which is not reverted with this commit.

rdar://79415891