[5.5] Fix constantness diagnostics for single expression closures
(This is a cherry-pick of PR #38053 to the release/5.5 branch.)

Due to the way miscellaneous diagnostics work with closure bodies, the
function argument constantness sema check does not produce proper
diagnostics for single-expression closures. This change fixes the
problem by having the ASTWalker manually walk in the closure body
if it is a single expression.

rdar://65577070