Fix TBDGen datalayout
This patch fixes the call to get the datalayout from the clang module.
Clang no longer holds onto the DataLayout object, so we have to do that
if we want to keep it around. Rather than instanciating it early, we can
just hold onto the string description, which clang does keep around, and
only construct the actual DataLayout once we need it.