[NFC-ish] Extract ASTNodeDumper
This is an RAII class that represents a single node (i.e. pair of parentheses in the S-expression output) in the AST dump. It automatically manages the indentation and printing of closing parens.

This commit converts PrintPattern, PrintDecl, PrintStmt, PrintExpr, and PrintTypeRepr to use it. We must further harmonize with PrintType’s semantics before it can be incorporated.

I’ve made minor changes to the way various nodes are printed, so tests have been updated to match.