[Demangler] Make isSpecialized() more robust.
We've had a couple of instances where isSpecialized() crashed while while trying
to use remote mirror to inspect a process that had itself crashed.  I can't see
any obvious mechanism for that to happen, so for now I decided to make
isSpecialized() more robust in the hope that if it happens again we'll get more
information and might be able to work out how it happened.

rdar://74394596