Sema: Typo correction should use GenericSignature and not GenericSignatureBuilder
Typo correction would call directly into the GenericSignatureBuilder
to get a list of all nested types of a type parameter.

Since the time that code was written, higher level APIs were added to
GenericSignature which originally wrapped the GenericSignatureBuilder.

These days when the rewrite system is enabled, the GenericSignature
operations also use the rewrite system, and the long-term goal is
to get rid of GenericSignatureBuilder altogether.