[Test] Add @escaping to async refactoring tests
The async refactorings previously ignored whether a completion handler
had `@escaping` or not. This is no longer the case, so fix up all
functions to have `@escaping` for their completion handler parameter.

Also some small miscellaneous fixes in order to reduce the number of
warnings output on test failures and also the addition of `REQUIRES:
concurrency` on all tests.