[TBDGen] Use default kind for AFPAST SILDeclRef.
When adding an AsyncFunctionPointer to the TBD, a LinkEntity of kind
AsyncFunctionPointerAST is constructed containing the decl.  That decl
is then wrapped in a SILDeclRef which is then mangled.

Previously, the kind of the SILDeclRef was erroneously forced to be
Func.  The result was a failure to mangle the TBD symbol for async
constructors correctly.

Here, that argument is omitted so that the kind can be determined
appropriately by SILDeclRef's constructor.

rdar://80485869