fix SILGen of access to isolated wrapper around instance actor property
The fix is to implement part of SILGen that I asserted could not
happen because I couldn't formulate a situation where an access
to instance-isolated state would happen at the point in the code.

But, if a property wrapper's wrapped-value is global-actor
isolated, and is wrapping an actor instance, then it does happen.

Implementation details:

If the address is isolated to an actor instance, then we need the
base in order to perform a hop prior to starting the access,
so I just pass the base value through SILGen to the point it is
needed, since the base is available from all of the callers.